fix strategy rewiring procedure
Previously, when i - a cooperator - cut a tie with j - a defector - and made a tie to a random agent k, k did not reciprocate this tie. This is a carryover problem from me (Elle) forgetting that we weren't working with directed ties in these simulations. Fixed and my bad